extract snapshot option
